Item images, NPC layout & Statuses, reorderable player moves
Adds the image editing icon back to all Item sheets, and disables the
(non-functional) attempt at setting default images for Items and Actors.

Moves the NPC Fatigue bar into the same left-hand column as their
Technique, to allow for a longer right-hand column that now includes a
first pass of Status tracking. (This still needs significant UX thought)

Renames the `filter_and_sort` method to `filter_items`, removing the
sort operation entirely, to allow the use of `item-list` for both Moves
and Techniques on the Player sheet, so that they can be manually
reordered to suit player preference.